Fix compilation error with various compilers:  - gcc 5.0, 6.0 and 7.4.0  - Clang 5.0

+++ examples/cpp/ponca_binding.cpp
@@ -118,7 +118,7 @@ void test_fit(Fit& _fit,
 // Build an interlaced array containing _n position and normal vectors
 Scalar* buildInterlacedArray(int _n)
 {
-    Scalar* interlacedArray = new Scalar[uint(2*DIMENSION*_n)];
+    Scalar* interlacedArray = new Scalar[uint8_t(2*DIMENSION*_n)];
 
     for(int k=0; k<_n; ++k)
     {
